[PT][SDK][Engine] Do not implicitly convert data types in `Model.execute` (#45283)
It is unclear why we allowed implicit data types conversions for tensor
inputs, since this will usually trigger an allocation. This was also not
applied for nested input types (e.g. list of tensors), so it was
inconsistent. This patch simply removes this behavior, and fixes tests
that relied on it.

MAX_INTEGRATION_TESTS_REV_ID: 9b5ee3192964b258a028d154c9093ca3b5bd0867

+++ max/tests/integration/API/python/pytests/test_telemetry.py
@@ -7,7 +7,6 @@
 import os
 import tempfile
 from pathlib import Path
-from subprocess import run
 
 import numpy as np
 from max.engine import InferenceSession
@@ -19,7 +18,7 @@ def test_api_source(mo_model_path: Path):
     os.environ["MODULAR_TELEMETRY_EXPORTERS_LOGS_FILE_PATH"] = filepath
     session = InferenceSession()
     model = session.load(mo_model_path)
-    output = model.execute(input=np.ones((5)))
+    output = model.execute(input=np.ones(5, dtype=np.float32))
     expected_line = "max.engine.api.language: python"
 
     with open(filepath, "r") as file:
